requirements.txt
aiohttp==3.6.2
astroid==2.3.3
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.8.2
certifi==2019.11.28
chardet==3.0.4
configparser==5.0.0
dice==3.1.0
discord.py==1.3.2
docopt==0.6.2
idna==2.9
idna-ssl==1.1.0
isort==4.3.21
lazy-object-proxy==1.4.3
lxml==4.5.0
mccabe==0.6.1
multidict==4.7.5
PyJWT==1.7.1
pylint==2.4.4
pymongo==3.10.1
pyparsing==2.4.7
pytz==2019.3
requests==2.23.0
six==1.14.0
soupsieve==2.0
twilio==6.38.0
typed-ast==1.4.1
typing-extensions==3.7.4.1
urllib3==1.25.8
websockets==8.1
wrapt==1.11.2
yarl==1.4.2
